backend/requirements.txt
Flask==3.0.3
urllib3==1.26.15
python-dotenv==1.0.1
langchain==0.3.7